V3 HiddenServices and Key Management

+++ tor/rootfs/etc/cont-init.d/52-tor-services.sh
@@ -12,8 +12,10 @@ declare clientname
 readonly torrc='/etc/tor/torrc'
 readonly hidden_service_dir='/ssl/tor/hidden_service'
 
+
 if bashio::config.true 'hidden_services'; then
     echo "HiddenServiceDir $hidden_service_dir" >> "$torrc"
+    
     for port in $(bashio::config 'ports'); do
         count=$(echo "${port}" | sed 's/[^:]//g'| awk '{ print length }')
         if [[ "${count}" == 0 ]]; then
@@ -50,33 +52,39 @@ if bashio::config.true 'hidden_services'; then
         # https://community.torproject.org/onion-services/advanced/client-auth/
         while read -r clientname; do
             # Generate a private key
-            openssl genpkey -algorithm x25519 -out /tmp/k1.prv.pem
-            # Format keys into base32, removing spaces.
-            # Using `sed` to prepare the string so we don't need
-            # the external tool (base64pem from basez) suggested
-            # in the documentation.
-            # Private key
-            cat /tmp/k1.prv.pem \
-                | sed -e '/----.*PRIVATE KEY----\|^[[:space:]]*$/d' \
-                | base64 -d \
-                | tail -c 32 \
-                | base32 \
-                | sed 's/=//g' > /tmp/k1.prv.key
-            # Public
-            openssl pkey -in /tmp/k1.prv.pem -pubout \
-                | sed -e '/----.*PUBLIC KEY----\|^[[:space:]]*$/d' \
-                | base64 -d \
-                | tail -c 32 \
-                | base32 \
-                | sed 's/=//g' > /tmp/k1.pub.key
-            # Create authorized client file
-            echo "descriptor:x25519:`cat /tmp/k1.pub.key`" \
-                > $hidden_service_dir/authorized_clients/${clientname}.auth
-            bashio::log.info "Private key for ${clientname}: `cat /tmp/k1.prv.key`"
-            # Delete the keys, as we don't need them anymore
-            rm /tmp/k1.prv.pem
-            rm /tmp/k1.prv.key
-            rm /tmp/k1.pub.key
+            # Only if client is a new one 
+            if [ ! -f $hidden_service_dir/authorized_clients/${clientname}.auth ]
+            then 
+                openssl genpkey -algorithm x25519 -out /tmp/k1.prv.pem
+                # Format keys into base32, removing spaces.
+                # Using `sed` to prepare the string so we don't need
+                # the external tool (base64pem from basez) suggested
+                # in the documentation.
+                # Private key
+                cat /tmp/k1.prv.pem \
+                    | sed -e '/----.*PRIVATE KEY----\|^[[:space:]]*$/d' \
+                    | base64 -d \
+                    | tail -c 32 \
+                    | base32 \
+                    | sed 's/=//g' > /tmp/k1.prv.key
+                # Public
+                openssl pkey -in /tmp/k1.prv.pem -pubout \
+                    | sed -e '/----.*PUBLIC KEY----\|^[[:space:]]*$/d' \
+                    | base64 -d \
+                    | tail -c 32 \
+                    | base32 \
+                    | sed 's/=//g' > /tmp/k1.pub.key
+                # Create authorized client file
+                echo "descriptor:x25519:`cat /tmp/k1.pub.key`" \
+                    > $hidden_service_dir/authorized_clients/${clientname}.auth
+                bashio::log.info "Private key for ${clientname}: `cat /tmp/k1.prv.key`"
+                # Delete the keys, as we don't need them anymore
+                rm /tmp/k1.prv.pem
+                rm /tmp/k1.prv.key
+                rm /tmp/k1.pub.key
+            else
+                bashio::log.info "Private key for ${clientname} already given"
+            fi
         done <<< "$(bashio::config 'client_names')"
     fi
